+++ randomgen/philox.pyx
@@ -207,7 +207,7 @@ cdef class Philox(BitGenerator):
     def __repr__(self):
         out = object.__repr__(self)
         out = out.replace("Philox",
-                          "Philox (" + str(self.n) + "x" + str(self.w) + ")")
+                          "Philox(" + str(self.n) + "x" + str(self.w) + ")")
         return out
 
     cdef _reset_state_variables(self):